utilisation classe

+++ classe/tank.cpp
@@ -9,11 +9,11 @@ int tank::get_y(){
     return y;
 }
 
-double tank::get_ori(){
+float tank::get_ori(){
     return orientation;
 }
 
-double tank::get_vit(){
+float tank::get_vit(){
     return vitesse;
 }
 
@@ -26,10 +26,10 @@ void tank::set_y(int new_y) {
     y = new_y;  // Modifie la position Y du tank
 }
 
-void tank::set_ori(double new_ori) {
+void tank::set_ori(float new_ori) {
     orientation = new_ori;  // Modifie l'orientation du tank
 }
 
-void tank::set_vit(double new_vit) {
+void tank::set_vit(float new_vit) {
     vitesse = new_vit;  // Modifie la vitesse du tank
 }
